+++ backend/app/services/user_service.py
@@ -1,0 +1,8 @@
+from app.repository.user_repository import UserRepository
+from app.services.base_service import BaseService
+
+
+class UserService(BaseService):
+  def __init__(self, user_repository: UserRepository):
+    self.user_repository = user_repository
+    super().__init__(user_repository)
